added a few casts (int) ti interpreter_c.c to avoid warnings
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@17314 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ MiniSystem/miniCompiler/runbeta.c
@@ -13,7 +13,7 @@ void main(int argc, char *argv[])
       printf("Cannot open file\n");
     }
   size =ftell(F);
-  printf("Runbeta: %i %s %i\n",argc,argv[1],size);
+  printf("Runbeta: %i %s %i\n",argc,argv[1],(int)size);
   fseek(F, 0, SEEK_SET);  //same as rewind(f);
   printf("Rewind\n");
   bc = malloc(size);
